+++ history.md
@@ -1,3 +1,6 @@
+### ToDo
+- InterfaceData的setData中关于setProp的调用更改为setData的调用
+
 
 ### 0.1.0
 - 基于complex-data简化逻辑，实现基本的功能